Moved the ListUtils.java file from commons, because it's only used by graphics.
