+++ src/husacct/graphics/task/layout/SortedNodeList.java
@@ -1,6 +1,6 @@
 package husacct.graphics.task.layout;
 
-import husacct.common.ListUtils;
+import husacct.graphics.ListUtils;
 
 import java.util.ArrayList;
 import java.util.List;
